gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
boto3==1.34.0